+++ src/mtf/dm/cms/hhs/gov/fixtures.ts
@@ -1,10 +1,10 @@
 import { test as base, createBdd } from 'playwright-bdd';
-import { TestScenarioContext } from './tests/demoUI/context/TestScenarioContext';
+import { TestScenarioContext } from './tests/dispensingEntityEnrollment/context/TestScenarioContext';
 
 // -- List of Page Object Class -- //
 import { HomePageSidePanel } from './tests/manufacturerEnrollment/pages/sidePanel';
-import { LoginPage } from './tests/shared/pages/loginPage';
-import { HomePage } from './tests/shared/pages/homePage';
+import { LoginPage } from './tests/common/pages/loginPage';
+import { HomePage } from './tests/common/pages/homePage';
 import { Page } from '@playwright/test';
 
 
@@ -64,4 +64,4 @@ export const test = base.extend<pageObjFixtures>({
 });
 
 // -- BDD Keyword Fixtures -- //
-export const { Given, When, Then } = createBdd(test);
+export const { Given, When, Then, Before, After, BeforeAll, BeforeScenario, BeforeWorker, AfterAll, AfterScenario, AfterWorker } = createBdd(test);
